requirements.txt
# Core dependencies with compatible version pinning for security updates
python-telegram-bot~=20.3
SQLAlchemy~=2.0.19
psycopg2-binary~=2.9.7
redis~=4.5.5
python-dotenv~=1.0.0

# Rate limiting for production
aiolimiter~=1.1.0

# Database migrations
alembic~=1.12.0

# Testing dependencies (optional, uncomment for CI/CD)
# pytest~=7.4.0
# pytest-asyncio~=0.21.0
# coverage~=7.3.0
